+++ user-components-x64/foo_jscript_panel3/samples/jsplaylist/main.js
@@ -1858,9 +1858,9 @@ function get_font() {
 	g_font_12_bold = js_font(name, 12, true);
 	g_font_20_bold = js_font(name, 20, true);
 
-	g_font_fluent_12 = js_font("Segoe Fluent Icons", 12);
-	g_font_fluent_20 = js_font("Segoe Fluent Icons", 20);
-	g_font_fluent_40 = js_font("Segoe Fluent Icons", 40);
+	g_font_fluent_12 = js_font(JSP3_ICON_FONT_NAME, 12);
+	g_font_fluent_20 = js_font(JSP3_ICON_FONT_NAME, 20);
+	g_font_fluent_40 = js_font(JSP3_ICON_FONT_NAME, 40);
 
 	g_font_group1 = js_font(name, 16, true);
 	g_font_group2 = js_font(name, 14);